requirements.txt
requests
psycopg2-binary
dbt-postgres
SQLAlchemy
python-dotenv
pandas